a first weird building version with templated simulaiton, soes not pass tests
